Switch names to First Last order
Names listed on this page are in "Last First" order, which makes
comparisons difficult. The "First Last" versions are on the individual
member pages, so we _could_ get them by also scraping those, but for now
it's simpler to reverse the names here. This could potentially go wrong
if someone has a name that doesn't follow this simple format, but we can
fix that if that ever arises (and haven't already switched to scraping
the individual member pages by then).